+++ src/App.js
@@ -1,10 +1,8 @@
 import React, { useState } from 'react';
-import logo from './logo.svg';
-import { Link, animateScroll as scroll } from 'react-scroll';
 import './App.css';
 import Header from './components/Header';
 import Navbar from './components/Navbar';
-import About from './components/About';
+import About from './components/about';
 import Skills from './components/Skills';
 import Projects from './components/Projects';
 import Contact from './components/Contact';
